requirements.txt
streamlit
speechrecognition
pyaudio
gtts
playsound
googletrans==4.0.0-rc1
deepl
python-dotenv